imgit/api/serializers.py
from django.contrib.auth.models import User
from imgit_app.models import Gallery, Image, Comment
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']


class GallerySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Gallery
        fields = ['title', 'author', 'created_on', 'modified_on']


class ImageSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Image
        fields = ['gallery_id', 'img', 'created_on']


class CommentSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Comment
        fields = ['created_on', 'author', 'gallery_id', 'text']
